ENGR00235193 - GPS Support for i.MX53
To enable the latest GPS sensor logic, a new directory is created
to house the date needed by the Orion binaries.  In testing,
addition of the GMS binaries required an increase in the system
image size.  Finally, although the gpsreset.sh file is not used
in the Orion configuration file, moved the binary to the correct
directory for Jelly Bean.

Signed-off-by: Jeff Kudrick <jeff.kudrick@freescale.com>